Move options parser
